+++ x/btcstaking/types/validate_parsed_message_test.go
@@ -891,7 +891,7 @@ func TestValidateParsedMessageAgainstTheParams(t *testing.T) {
 
 			msg, params, _ := tt.fn(r, t)
 
-			parsed, err := types.ParseCreateDelegationMessage(msg)
+			parsed, err := msg.ToParsed()
 
 			if tt.errParsing != nil {
 				require.Error(t, err)